Made Two-Column Code Much Smaller
The code used for two-column figure series is now much smaller. I cut
off a lot of rather useless stuff. The "Float(s) lost" problem is not
yet solved. However, with much smaller code, I think my chances have
increased to find the vital spot causing the problems. Due to the
shortening of the code, the documentation is now out of sync with the
code. I will fix this later.